Ceph: configure the ceph::osd::excluded_slots per cluster
Each cephosd cluster member has a pair of disks that are designated for
use as the operating system drive and are therefore excluded from the
list of disks for which an OSD is created.

For the cluster in eqiad, these are in the enclosure with the number 23,
but in codfw the enclosure is numbered 25.

Hosts: cephosd1001.eqiad.wmnet,cephosd2001.codfw.wmnet
Bug: T374923
Change-Id: I8bb143e3de6427ace8b60fa0b455dfe1e146ea38