ceph::osds - Use sdparm instead of hdparm to disable the write cache
When commissioning the cephosd cluster on codfw, it was observed that a
number of the SSD drives give an error in response to any hdparm
commands, which attempt to disable the write cache.

In order to address this, we switch the ceph::osds class to use the
sdparm equivalent commands and to install this utility instead of
hdparm.

Hosts: cephosd1001.eqiad.wmnet,cephosd2001.codfw.wmnet
Bug: T374923
Change-Id: If47852abcc4d21bc42b8ba9c7294b30f16ad7e45