Cephosd: add hiera keys for rack-specific BGP peers in codfw
The Bird Anycast profile defaults to configuring the core router
loopback IPs in a given site as the BGP peers in the bird config.

As these hosts in codfw are connected to L3 switches the hosts
instead need to peer with their top-of-rack switch on the local
GW IP.

Adding these to force the correct ones.

Bug: T374923
Change-Id: I9f40ea3b22fb7368f61ec063c1b7838f8ff0cbfe