reimage: use ipxe DHCP info, skip d-i DHCP
Debian's DHCP client, udhcpc, does not include option 97, pxe-client-id,
or the machine UUID with its DHCP discover packet. This causes DHCP to
fail if the DHCP server is configured to match based on the UUID of the
DHCP client. We could possibly teach the debian installer to add option
97 with udhcpc's -x option, but since we already have all the info in
ipxe, why not skip the DHCP step entirely? The below config was tested
successfully with sretest2001.

1. http://blog.loftninjas.org/2008/07/17/booting-debian-installer-with-a-static-ip-address
2. https://gist.github.com/tuxfight3r/76827b554443260c9410b5196bbc5148

Hosts: O:apt_repo
Change-Id: I8e194bc0a2a349cd7463c3c0d9c8b83fa42aefc2