late-command: Check whether qemu_fw_cfg.ko is present
For releases in flux like the current Ganeti, the kernel ABI change differ
between the kernel d-i was booted with (currently 6.12.33+deb13-amd64)
and the udebs being installed d-i itself (currently 6.12.35+deb13-amd64).

As such, test whether qemu_fw_cfg.ko is actually present, otherwise
the installation fails. This won't work with VMs on routed Ganeti, but
for the interim tests of trixie that's not really relevant.

Bug: T391083
Change-Id: Iad08ee80bfe4ced5b6d546a2b1937fd4a2ee3622