profile::kubernetes::mediawiki_runner: add feature_flag support
When manipulating Puppet code to test or roll out
experimental features, it is not always clear that
a specific piece of code is part of an experiment.

We can improve the visibility of such experiments by
introducing a generic feature flag, we can adapt at will.

Hosts: wikikube-worker1300.eqiad.wmnet
Change-Id: Iea6721b79ee9f70b135a5053119f057cf6f0ff7e