k8s::mediawiki_runner: allow outgoing connections to memcached
We have established that a fair number of MediaWiki memcached
errors are produced due to the reloading of ferm rules. During the
timeframe between a ferm reload and kube-proxy reapplying the
correct firewall rules, MediaWiki pods are unable to access memcached.

While this solution is not perfect, it will improve the situation.

Hosts: wikikube-worker1300.eqiad.wmnet
Bug: T371881
Change-Id: Icc0f0cfb8832809df6bb64baa015e1831b371718