Analytics: Refine post migration update
In order to have a clear cut at 7/15 00:00 of the
Refine migration. We reduced temporarly the
time range of the legacy Refine.

We miss to set a smaller time range for the
monitor job too.

Now that the migration is done. We increase the
time range to make it refine on 24 hours as
before.

And to avoid monitoring alerts, we monitor
only after the clear cut time. Currently ~32h.
Later we will set it back to 48h.

Bug: T369845
Change-Id: I42c02bfd774a05b0af1a1c9c1c53de4631c725ff