Tweak the java options for hive-metastore on the test cluster
We have seen performance problems with regard to the hive-metastore
recently, as its usage has increased significantly.

As a result, we wish to test these changes to the java options on the
test cluster, then promote them to production if they function
correctly.

* Reduce MaxGCPauseMillis from 1000ms to 200ms
* Reduce InitiatingHeapOccupancyPercent from 45 to 30 to start marking
  earlier
* Set G1ReservePercent to 15 to avoid long allocation stalls

The effect (when in production) should be to increase the CPU slightly,
as more frequent GCs will be taking place, but this should result in
overall lower latency for high-throughput metastore access.

Hosts: O:analytics_test_cluster::coordinator
Hosts: O:analytics_cluster::coordinator
Bug: T399711
Change-Id: I1b7436cbe6bd5b64271726e0746199b94e86461c