Apply the hive-metastore GC changes to production
We have tested three changes to the hive-metastore GC settings in
staging, so now we wish to apply them to production.

* Reduce MaxGCPauseMillis from 1000ms to 200ms
* Reduce InitiatingHeapOccupancyPercent from 45 to 30 to start marking
  earlier
* Set G1ReservePercent to 15 to avoid long allocation stalls

The effect (when in production) should be to increase the CPU slightly,
as more frequent GCs will be taking place, but this should result in
overall lower latency for high-throughput metastore access.

Hosts: O:analytics_test_cluster::coordinator
Hosts: O:analytics_cluster::coordinator
Bug: T399711
Change-Id: Ibb4c7ff71bd9d407be8078eb7672dffacba1f3e7