cephosd: un-set bird bgp neighbors rather than override for each host
I did not properly look at the bird role previously when we added
these.  By default the bird role will get our core router loopback
IPs for the particular site in profile::bird::neighbors_list.

As the cephosd's are all in racks with routed/layer-3 switches we
instead want to peer with the top-of-rack switch IP.  As such I
added host files for each with the correct IP.

It turns out the role will use it's default gateway as neighbor if
the var is unset, so we can instead just do that for the role and
avoid having to list out the IPs which is more robust.

Hosts: cephosd1001.eqiad.wmnet, cephosd1005.eqiad.wmnet
Hosts: cephosd2001.codfw.wmnet, cephosd2003.codfw.wmnet
Change-Id: I6a560ec285af31ee2c7e5b881020061faced157e