pyrra: simplify multi-dc handling for istio SLOs
We are considering Istio services behind a discovery service,
and it seems good to duplicate error budgets by SLO. Using a regex
match like =~ for "site" triggers (in Pyrra) an automatic grouping in
the generated recording rules, that is not what we want.

While at it, fix the k8s cluster name for tone check.

Bug: T398534
Change-Id: Idc82c8a4861126b23d75c6c210b7513054ec4ccb