Bigtop: Move the excluded_hosts values into profile::hadoop::common
When setting Hadoop nodes into decommissioning mode, we have until now
been adding their hostnames to specific hiera values that are only
relevant to the Hadoop master and standby master hosts.

What this meant in practice was that the hadoop-yarn-nodemanager service
would fail to start on all of the excluded hosts, with subsequent alerts
being generated from systemd and from the puppet-agent.

This change moves the list of excluded_hosts into the hadoop_clusters
hiera structure that is defined in common.yaml and the value can then be
looked up by the workers. If a worker finds that it in the list of
excluded hosts, then it will stop the hadoop-yarn-hdfs service and this
will stop the alerts from being generated.

Hosts: O:analytics_cluster::hadoop::master
Hosts: O:analytics_cluster::hadoop::standby
Hosts: an-worker1078.eqiad.wmnet  # present in excluded_hosts
Hosts: an-worker1207.eqiad.wmnet  # absent from excluded_hosts
Bug: T397160
Change-Id: Id207907398649581424f461121b00aef3c23729d