gitlab: exclude packages from failover backup
This change excludes packages from the failover backup. Packages contain
CI generated files with more than 200GB+ of file size. To reduce
the downtime during failovers packages are excluded. The estimated
downtime is reduced from 5h+ to less than 2h.

Worst case 24h of packages are lost and have to be re-created. This will
be fixed when packages are moved to object storage (similar to "artifacts").

Bug: T399306
Change-Id: Ice9f6374bbeb7ae81b815a0bfbf2703314ee4a81