Fix an issue with disabling the hadoop-yarn-nodemanager service
We recently deployed a patch that was intended to stop the
hadoop-yarn-nodemanager service running on hosts that are in
'decommissioning' mode.

It didn't have the intended outcome, because the systemd::service
resource was also an override. The result was that we removed the
override, while continuning to run the service.

This patch corrects that issue by reinstating the override and passing
the 'ensure' parameter as part of the 'service_params' has instead.

Hosts: an-worker1078.eqiad.wmnet  # present in excluded_hosts
Hosts: an-worker1207.eqiad.wmnet  # absent from excluded_hosts
Bug: T397160
Change-Id: I8ecb25175bae7cbbb07e1649819ccde8c0cadc99