opensearch: curator instance config to follow $enable_curator
Why: The opensearch::curator::config define requires
opensearch::curator to be included, but it isn't when $enable_curator
is false.

Later, opensearch::instance define attempts to create the curator
configuration, but fails because the path does not exist on the host.

What: Gate provisioning of the config on $enable_curator so that the
instance config does not fail because /etc/curator does not exit.

Hosts: logstash1023.eqiad.wmnet,logging-sd1001.eqiad.wmnet
Bug: T353912
Change-Id: Iaa79e1d84297e3c6a5967c91c4015768637b04a2