trafficserver: Add fractional routing to gateway-check
Why:
In order to progressively migrate API routes to a different backend, we
need a fractional routing facility in the routing script.

What:
Adapt what was already done for mw-on-k8s and mw-php-migration and add a
third field to the rules config that sets what fraction of the traffic
should be using the rule.

Hosts: P:trafficserver::backend
Bug: T400131
Change-Id: Ica40926671066ef62bda18bf5ac21c1710831b40