wikireplicas: maintain-views: Allow running without replacing views
Replaces the --replace-all flag, which skipped the confirmation prompt
for replacing views, with a new --replace flag, which will allow
replacing views (without confirmation). If the new flag is not set, no
views will be replaced.

This is designed to help operations like dropping views to avoid having
to drop and re-create all the views at the same time, or being prompted
about it for each view.

Change-Id: I6a6a6964572b6d1adba381b2b09b4e9c04ffe425