install_server: fix hwraid-1dev-nvme and modify boss_leavelvm.cfg
The sretest2006 host has a Dell Boss card with two nvmes in raid1,
exposed as single device. My previous attempts to create a generic
recipe using standard-efi.cfg were a failure since, among the other
things, a RAID device is not needed.

I noticed the boss_leavelvm.cfg recipe, used for apus and cephosd
nodes, so I thought to refactor things in the following way:
- Move any non-ceph specific bit into hwraid1-1dev-nvme.cfg.
- Leave out only the LVM specific bits as "custom".

I suspect that we could simplify even further setting the remove lvm
feature as true in hwraid1-1dev-nvme.cfg, and override it only
in boss_leavelvm.cfg, but I am not that confident in partman to
risk ceph's volumes while reimaging.

Bug: T393044
Change-Id: Ic32502b99e8212d24437ba4de9af4fc980db49ce