profile::thanos::recording_rules: add two rules for the EditCheck SLO
The Editing team have created an SLI related to a simple comparison
between mediawiki_editcheck_preSaveChecks_total time series, and
in Pyrra we'd need to add the following ratio:

(mediawiki_editcheck_preSaveChecks_total{kind="Available"} -
  ignoring(kind) mediawiki_editcheck_preSaveChecks_total{kind="NotShown"} -
  ignoring(kind) mediawiki_editcheck_preSaveChecks_total{kind="Shown"})
    / ignoring(kind) mediawiki_editcheck_preSaveChecks_total{kind="Available"}

So the idea is to have an SLI that measures the skew between the sum
of NotShown and Shown with Available. In Pyrra we cannot set
the aforementioned formula directly, because we need to provide two
inputs: error and total. I am adding recoring rules for both use cases
since, as far as I can see, it seems difficult to apply increase()
to the result of a binary operator (in this specific case not all labels
are matching and this is why I had to put 'ignoring(kind)').

Pyrra will in fact generate its own recording rules using increase() or rate().

The total use case looks a bit weird since I had to find a way to
drop the kind label, so that the resulting metric could be used
in a binary operation like "-" with the other one (there may be
a smarter way to do it).

Bug: T395444
Change-Id: I5518bc9cf174522e7e12f2244b5cc45ded96a9dc