zuul: add mysql prod password in new zuul config
The password to the mysql db created in T394844
existed in the private puppet repository but needs to be
added to the new zuul.conf created from an erb template.

In production we will not be using docker compose, like
in the test environment. So do not include the password
from there and remove the related comments.

Bug: T395938
Change-Id: I26d09bbcf0ccc15a6a1c8f34bb7eb96d8c7925cd