proxy: Allow outbound HTTPS connections to port 6443
Update the HTTP proxy squid service configuration to allow outbound HTTPS
connections to remote services hosted on port 6443. This is the port
that the Kubernetes API listens to by default. Opening it as
a connection target will allow the production hosted nodepool instance
used by zuul to talk to a Kubernetes cluster hosted in Cloud VPS.

Bug: T394838
Change-Id: I799566637421a2e3660d5e97fa7c0adb89ef92fc