zuul: use mariadb connector and Hiera'ize mysql_host name
zuul docs say:

"Zuul supports PostgreSQL, MySQL, and MariaDB. Supported SQLAlchemy
dialects and drivers are .." and:

"If using MariaDB, be sure to use the mariadb dialect."

https://zuul-ci.org/docs/zuul/latest/configuration.html#database

We are using MariaDB in production so do what it says and use that
dialect.

https://docs.sqlalchemy.org/en/20/core/engines.html#database-urls

https://docs.sqlalchemy.org/en/20/dialects/mysql.html#module-sqlalchemy.dialects.mysql.mariadbconnector

Also, puppetize the mysql host name and put the value in Hiera.

We are using m1-master.eqiad.wmnet per T394844#10894356.

Bug: T395938
Change-Id: I1025bc8d177c599feabf5872713140ff87abdb1d