mediawiki: Fix non-redirecting download.mediawiki.org domain
This is exposing various unintended entrypoints from here, essentially
treating it as a complete wiki, despite having no history of this
domain being its own wiki. Such as:

* https://download.mediawiki.org/wiki/Special:Blankpage
* https://download.mediawiki.org/w/api.php
* https://download.mediawiki.org/robots.txt

etc

Apart from the plain https://download.mediawiki.org URL, most
other paths did not redirect, unless Apache or MediaWiki would already
redirecting for a different reason, in which case (thanks to
UseCanonicalName) it will indeed redirect to the canonical name at
the same time in that case. By itself, UseCanonicalName is not a
redirect instruction.

Use a funnel instead for the domain as a whole.

While at it, point it to a more exact destination as well, rather
than the generic mediawiki.org homepage.

Bug: T50517
Bug: T80222
Change-Id: I9963688a805828ea976bbce42a9fdf28b6b3d893