gerrit: replicate repo renames as "gerrit2" application user
While adding the rename-project plugin replication config in
In I97fe0719da5923542ae0add11a0831e9e728d157, I have erroneously
requested to use "gerrit" as a user to connect to the Gerrit daemon on
gerrit2003.

Although the daemon is using "gerrit" as a Unix user, there is no such
user in the Gerrit database and a ssh results to the daemon yields:

  [SSHD] gerrit - AUTH FAILURE FROM 208.80.154.135 - - - user-not-found - - - -

Change the user to `gerrit2` which is in the `Administrators` group.

Bug: T239693
Bug: T398401
Change-Id: I776d0f80183360f535e88ecf294e1f861a3e7752