thanos: drain thanos-be1005 for disk controller swap
This node and thanos-be2005 need a replacement disk controller (that
supports JBOD properly); once that's done the disk devices are
differently named. So drain the node, then remove it from the rings,
swap the disk controller, and add it back with the new device names.

This is the first part of that process.

Bug: T400877
Bug: T400878
Change-Id: Ice7a82369876331cdbcdf520d0ee5a0822243741