Begin cleanup up the clouddumps servers
We are decommissioning all of the snapshot and dumpsdata servers, but
the clouddumps servers are remaining. This commit absents a few
resources that refer to the old servers from the clouddumps servers.

We will then remove the puppet close itself in a subsequent commit.

Hosts: clouddumps1001.wikimedia.org,clouddumps1002.wikimedia.org
Bug: T398438
Change-Id: I3fdfcfb225a19c21a7a98e1e552ff1c87afe7923