Remove all puppet code related to snapshot and dumpsdata hosts
We have now migrated all of the dumps v1 processes to Airflow and
Kubernetes, and the snapshot and dumpsdata hosts have been prepared for
decommissioning.

This commit removes all of the remaining puppet code related to these
roles that are no longer used.

Hosts: clouddumps1002.wikimedia.org,clouddumps1001.wikimedia.org
Bug: T398438
Change-Id: Ia177255f1385c6dd8b810939bc08b82bcc1baf23