varnish: refactor inclusion of requestctl rules
Instead of having the two different templates with different logic for
cache hits and cache misses, include both files in the cache miss case.

This way, we can refactor the confd perversion to a single file that
should work generally for all scopes.

This is the first step in rationalizing how we manage requestctl rules with
different scopes in varnish.

Hosts: cp1100.eqiad.wmnet,cp1101.eqiad.wmnet
Bug: T399058
Change-Id: Ia4e08ce656885918f84e8c62aed8d8ab7109d170