gNMI: initial Nokia support
The `interfaces-states` path is OpenConfig, so OpenConfig needs
to be enabled on the Nokia device with:
`set system management openconfig admin-state enable`
`prefix` tested on Juniper as well.
The alternative is to set the default model in nokia as openconfig with:
`set system grpc-server mgmt yang-models openconfig`

In that case we need to add the prefix `native:` when querying Nokia
specific paths.

Hosts: netflow2003.codfw.wmnet
Change-Id: I7e3619bc1ef878aaee6c65a1058e5b2759230855