Add more dse-k8s-worker nodes to site.pp
We are re-commissioning all of the previous snapshot servers as
dse-k8s-worker nodes, having completed the migration of Dumps to
Airflow.

The first step is to add the new hostnames to site.pp in the correct
insetup role. They will match the correct partman recipe in preseed.yaml
already.

We will remove the references to snapshot hosts in a sunsequent commit,
once all of the renames are done.

Bug: T398438
Change-Id: Id1e8e96ed9c1aae625e4006c7e8057bb8025a6b9