swift: remove old nodes, drain & reweight SM C-J nodes
A set of related changes to the eqiad ms rings:

1) remove ms-be16[1-3] (now at 0 weight) for decom
2) create a schema prod24_sm_jbod to reflect the device names on
SuperMicro Config-J nodes with the new JBOD card in
3) add ms-be1091 to the rings with this new schema
4) drain 3 SM C-J nodes so they can have their controller swapped

We are currently using 3.685P raw storage, making all the changes
above reduces our raw capacity to 5.568P, meaning we'd be 66% full,
which is comfotable. There are 9 SM C-J nodes to do in eqiad, so doing
them in batches of 3 makes sense.

Bug: T391354
Bug: T401127
Bug: T400877
Change-Id: I752a8606c86fa640eb9c6cc181bc4742abcd35e5