admin: upgrade user osleger, add to parsoid-test-roots
Otto Šleger needs a shell user as requested in T401300
and should then be added to the group parsoid-test-roots.

He already has LDAP access so it's an upgrade to shell
user. LDAP groups are unchanged by this.

For more context on groups see T401300#11065770.

Approval by manager and group owner is already on ticket.

Bug: T401300
Change-Id: I38d6d310337be355d4bd111cdd50b3f86e203bb9