swift: add 1 new codfw host, drain 3
We're adding ms-be2088, which is a host with the new JBOD disk
controller, so needs the new storage schema (like ms-be1091 in
eqiad).

Additionally, we are draining 3 nodes prior to reimage - ms-be2079 so
it can be moved to the new per-rack network (T354872) and ms-be2081/2
so they can have the new JBOD disk controller put in (T400876).

Bug: T382056
Bug: T354872
Bug: T400876
Change-Id: I0e8fd953fc449cf804e18d6a935b5bb5503e34cf