admin: allow systemctl status for MediaWiki train
We have systemd units to manage the MediaWiki train and sudo rules to
start and inspect the journal.

Add rules to allow `systemctl status` which is a short way to get the
status of the unit AND the last few lines of the journal.

Without sudo, we do not have access to the journal. The status
subcommand does not show the logs:

  Warning: some journal files were not opened due to insufficient
  permissions.

Follows-up to b08fc1f62d

Change-Id: Iafd96511466bf8ed0ef29dabb4c946ab669c3edc