Add the new an-backup-namenode hosts to puppet
These hosts were formerly an-worker1100 and an-worker1101 which were
Hadoop workers with 24 x 1.8TB drives and GPUs.

Now they will be reused as namenodes for the HDFS backup cluster, so we
will reconfigure the RAID array to make them have one large RAID10
/srv volume instead of 24 x RAID0 volumes.

This commit adds them to the site.pp and preseed.yaml files, along with
a suitable partman recipe.

Bug: T397175
Change-Id: Ic6b40091e9e749d04b89dfa70f3949bbd36513ce