zuul: create empty dir /var/lib/zuul on new zuul main hosts
Ensure /var/lib/zuul exists as an empty dir / mount point on
new zuul main hosts.

So that it can be used as a mount point in the docker command
as an option to avoid docker volumes.

This directory is the home directory for the zuul user in the
image.

Bug: T395938
Change-Id: Ic3450281f7204a7cb7f9cd3b4fa9733ae5b59c40