mariadb: Remove catalogued tables from private_tables in hiera
Those tables are already present in the table catalog in
modules/mediawiki/files/mariadb/tables-catalog.yaml, and no
longer need to be present in private_tables.

This keeps the `cur` and `hidden` tables, as those are marked
as dropped and I do not know whether that would be sufficient.

Bug: T399302
Hosts: db1154.eqiad.wmnet
Change-Id: I1b4096c159f9d4723476726311626a2ff02c0a4e