openstack: vendordata: Use puppet-agent as Puppet package
Install and possible downgrade the puppet-agent package, not the
'puppet' transitional package. The puppet::agent class is still going to
install the puppet package, that will need to be fixed in a follow-up.
But that happens late enough that the repos/pins for the Puppet 7
forward port should already be in place.

This might break installing new Bullseye base images, but I think that
is fine.

Bug: T401913
Change-Id: I6a6a69642bbca1637c09da222e8621d4828fb1bf