mediawiki: Remove unused wikidata.org vhost and fix beta redirect
== Why ==

Prod has a ServerAlias in the Apache conf of www.wikidata.org for
apex wikidata.org, which seemed like it was responsible for the
the redirect to www, and so in T289318 I copied that to Beta.

In actuality, ServerAlias+UseCanonicalName do not produce a redirect,
it merely rewrites the request so that MediaWiki sees it as if it was
for the canonical www.wikidata.org and then responds in-place.
This means we don't redirect these:
* https://wikidata.beta.wmcloud.org/wiki/Wikidata:Main_Page
* https://wikidata.beta.wmcloud.org/w/api.php
* etc

The exception here is cases where MediaWiki would already redirect
regardless (such as "/" to "/wiki/Main_Page" which also set the
canonical domain while at it).

It is undesirable to expose MediaWiki under non-standard vhosts:
* it creates a mismatch in self-identity (e.g. JavaScript errors due
  to references to the "same" domain becoming cross-origin references
  and thus being denied). Example reported in T401592.
* exposes a non-redirecting and un-purged stale copy under a
  non-canonical URL, which can be used in vandalism or phishing.

In production this was not an issue, because there is *also* an entry
in redirects.dat redirecting wikidata.org to www.wikidata.org, which
hid this construct from the public, with no other DNS entires under
the *.wikidata.org wildcard existing besides apex, www, and m.

== What ==

* Add a matching redirect for non-www wikidata to Beta.
* Remove unreachable vhost from prod.
* Remove unreachable vhost from Beta.

Once these three changes are applied, we find another piece of
tech debt that starts causing issues:

The Varnish logic for rewriting m-dot to canonical hardcodes the
production domains that use www and uses a generic m-dot stripping
for the rest (since other sites indeed correclty "find" the canonical
domain by merely stripping m-dot, whereas for www we must swap).

This meant that until now, it was actually sending traffic for
m.wikidata.beta.wmcloud to MediaWiki appservers as "wikidata.beta.wmcloud"
instead of "www." which then happened to work because of the
remnant wildcard vhost, which this patch removes.

* Fix Varnish logic to consider top_domain so that we correctly
  rewrite `m.wikidata.beta` to `www.wikidata.beta`.

Hosts: mwdebug1001.eqiad.wmnet,cp3070.esams.wmnet
Bug: T401592
Change-Id: I977e1f080405ca6760222e6e754bee3b05054446