bird::anycast: Add a parameter to install the Bird enabled for routed Ganeti
Support for using Bird/BGP with routed Ganeti was recently added by cz.nic.
Enabling this options enables a repository component with a build of Bird
based on a branch with that work. Eventually the release of Bird 2.18 will
include all this work (and when that has happened, the component updated).
If all roles using Bird are fully backwards compatible we can also migrate
them one by one and then finally update bird in the "main" component of
apt.wikimedia.org to 2.18

Hosts: C:profile::bird::anycast
Bug: T362392
Change-Id: I5dcf6f5d1103742ca790798d3ac32ed7f4e0f62f