cache::text: set apt-staging to NOT cache
Setting the "pass" parameter to prevent the CDN from
caching requests to apt-staging.wikimedia.org.

The reason is described on the ticket but asked for a configuration
of the maximum cache time on the nginx backend.

This option simply disables caching entirely while we keep the
other benefits of being behind the CDN.

Bug: T402284
Change-Id: Ief4475c4d858a4438c0a717d65edf0fd5d490dce