doc.wikimedia.org CSP: Allow sendBeacon for piwik (Matomo)
We were already allowing scripts to be loaded from piwik.wikimedia.org,
but that's not enough, we also need to allow `navigator.sendBeacon()` to
send data to that domain, and that requires `connect-src` in addition to
`script-src`.
While the sub-domain was chosen for the software's original name Piwik,
it is nowadays called Matomo.

Bug: T368970
Change-Id: I56d8487187ec51f5a4243687b87ed7b733503243