varnish: Implement new direct routing for mobile views
This untangles the distinction between a "mobile client"
and a request that is "redirect-able".

Consider:

* A mobile client visits https://en.wikipedia.org/wiki/Banana
* We serve the mobile view, by redirecting to
  https://en.m.wikipedia.org/wiki/Banana
* They perform a non-pageview action such as:
  POST https://en.m.wikipedia.org/w/index.php?action=submit
  GET https://en.m.wikipedia.org/w/index.php?mobileaction=toggle

  These non-pageview actions are performed within the mobile site,
  and thus use the mobile skin and layout in MediaWiki
  (via Minerva+MobileFrontend) as per the X-Subdomain:M header that
  is applied to the domain as a whole, not conditional on whether
  it is a pageview.

Previously, this function thus only dealt with redirects and bundled
these two concepts together. Once untangled, we first recognise
whether the client is considered a mobile client (which determines
whether we want them to have the mobile version), and then separately
(in the redirect branch) decide whether the current request qualifies
for such a redirect. The new branch will not have these conditions.

This way, the resulting experience is thus unchanged:

* Mobile visits https://test.wikipedia.org/wiki/Banana
* We serve the mobile view, by responding in-place using X-Subdomain:M.
* They perform a non-pageview action such as:
  POST https://test.wikipedia.org/w/index.php?action=submit
  GET https://test.wikipedia.org/w/index.php?mobileaction=toggle

  These must not suddenly expose the desktop site, but remain
  in the mobile experience.

Bug: T401595
Change-Id: I44108c0aed01dfcd8a200323a6d00b84d4a692aa