Add deprecations to varnish
We want to be able to deprecate urls via requestctl. When we do so,
we want these urls to be deprecated for everyone, including internal
clients and trusted sources.

To this end, add the specifically-scoped rules early in the process,
in cluster_fe_recv.

Hosts: cp6001.drmrs.wmnet,cp6016.drmrs.wmnet
Bug: T398161
Change-Id: I8553991e419f604585d812db2ce66c9a05a4e764