trafficserver: Check x-dt-host in rb-mw-mangling.lua before using
In I44108c0aed01dfcd8a2, we implememnt a new direct routing mode
for mobile pageviews in VCL. After that patch, the MobileFrontend
header (X-Subdomain:M, to be renamed to X-Mobile:1 some day) can
exist on both m-dot and on canonical requests.

This mangling code should only assign Host for m-dot reqs,
because the x-dt-host field does not exist on other requests.

Without this, ATS ends up calling MediaWiki on mobile-dr pageviews
with `Host: nil`, which is (unsurprisingly) not a recognised vhost
on MediaWiki's Apache server.

Bug: T401595
Change-Id: If23fd901989ab915fbd671747cc1223a1397a325