admin: add user aude to analytics-privatedata-users
aude is an existing shell (deployment) user and an
engineer in the Reader Experience team.

but per the request on the ticket:

"access to some dashboards in superset (e.g. xLab experiments) is
restricted"

Since it just affects some dashboards the difference will be
between dashboards with our without private data.

To enable access to the ones with private data the additional
analytics-privatedata-users group is needed.

reference:

https://wikitech.wikimedia.org/wiki/Data_Platform/Data_access#Dashboards_in_web_tools_like_Turnilo_and/or_Superset_that_do_not_access_private_data

Bug: T402022
Change-Id: I0cd3c0b46b1d7adef2e0646947fac1eecf635e82