mariadb: replace legacy fact for memorysize
For one reason or another this stops working and is not
set on trixie, causing puppet errors.

We use the same puppet version as before but the ruby version
changed. Legacy facts _should_ still work but somehow it's not
in this case.

I could confirm through local testing this is behaving differently
between bookworm and trixie.

Even though I can't fully explain all details it should be the right
thing regardless; since it is replacing a legacy fact with the more
modern version of it.

And the compiler output shows no diff:

https://puppet-compiler.wmflabs.org/output/1180999/6698/

Ideally this also unblocks using the template on any trixie
host.

My main reason to do it is to unblock upgrading my cloud-VPS-project
not production; but of course it touches this default template.

Change-Id: Id249ca4be0a11940a2db707681cb69fb88dbab32