benthos: Use kafka_offset to sample webrequest_live messages
Under certain circumstances HAProxy reports a sequence number of 0,
which bypasses the current sampling logic. This causes BADREQ requests
to be over-represented in Turnilo.

We intend to fix the underlying issue in
I85067eef56068943d2e419ada55808815fa0668e, but in the meantime switch
sampling to use `kafka_offset` to ensure consistent distribution of
webrequest_live messages.

Bug: T401383
Change-Id: I618158ed0ec935e2b63e31719a7fc503eea9bc21