Keystone hooks: speed up domain creation
Previously designatemakedomain waited for a domain to be active before
transferring it to the requesting project. This is reliable but slow,
often causing project creation to time out.

If we don't wait for things to be active, we can complete
project creation much more quickly. The cost is that domains
will show as 'pending' for several minutes, until designate
does some routine refreshing of the dns servers.

Bug: T398712
Change-Id: I42faa311b27c9c29068ba50d404d9a370b91108f