varnish: Remove 60s cap for mobileaction/useformat on m-dot
== Background ==

This traces back to 2011 in commit 6dd4620e15 (I1e149a215d3a0c),
for the purposes of avoiding serving stale content on these
URL variants because they are not canonical and thus never purged.

https://en.wikipedia.org/wiki/Main_Page?useformat=mobile

The "useformat" query parameter in MobileFrontend is similar to
"uselang" and "useskin" (in MediaWiki core) and mostly developer facing
as a way to manually enable MF mode on just one request.

This parameter works on both m-dot and canonical domains but for
some reason this TTL cap was only added in VCL for m-dot requests
(originally added to VCL files specific to the cache-mobile cluster,
and later moved here under on X-Subdomain to logically preserve that.)

This has not caused issues stale responses for the above URL,
because, MediaWiki historically responded without public Cache-Control
on non-canonical URLs. So, these did not have a long expiry anyway.

More recently, with T309063 in 2022, MediaWiki introduced cache
allowance of 1 hour for any ad-hoc URL variations such as this one.

https://en.m.wikipedia.org/w/index.php?title=Main_Page&mobileaction=toggle_view_mobile

The "mobileaction" has, at least for the past ten years, responded
with a deterministic 301 redirect that simply strips the parameter,
and thus no content to be stale.

This parameter too exists also in the reverse direction:

https://en.wikipedia.org/w/index.php?title=Main_Page&mobileaction=toggle_view_desktop

MediaWiki must set a cookie (stopMobileRedirect) here to avoid redirecting
subsequent pageviews to m-dot from this mobile client.

For the past ten years (since 2016 with T127993), MediaWiki
automatically marks set-cookie responses as uncachable (CC:private,
Expires:1970).

== What ==

Remove the ttl cap for these URLs.

== Why ==

It is redundant, because:

* useformat is already capped to 1h in MediaWiki which is enough.
  We do the same for uselang/useskin, and any other non-standard URL.
  We do the same for useformat itself, on canonical domain, because
  the logic was limited to m-dot for some reason.

* mobile opt-in toggle, is a content-less 301 redirect where purging
  is no concern, and MW further limits it to 1h per above.

I want to remove it, because as part of m-dot sunsetting, these
will all happen on the canonical domain going forward. This means
unless we refactor this to apply to equivalent unified URLs, the
code will become dead anyway, and I'd rather remove it explicitly
confirm its non-impact in isolation.

Bug: T401595
Change-Id: Ied2fbc954377510b6f73431c28b075f137f85070