beta: Update hieradata for fe_vcl_config from Horizon
Put the latest here so that this data is actually used. This makes
cherry-picking other Varnish patches easier, as it doesn't require
an out-of-bound Horizon edit that way, and makes it less likely for
SRE to break Beta as they can update these at the same time as
the production Hiera data files.

The config is identical between "text" (deployment-cache-text)
and "upload" (deployment-cache-upload) in Beta Cluster, except:
* `block_help` which overidden but not synced here,
* `measure_domain_regex` is nowadays required and was missing here,
* `purge_host_regex` varied between them (maps/upload-only vs anything
   except), which doesn't seem important in Beta so I replaced
  that by a wildcard match to keep it simple.

Change-Id: I55fcc0d98975615cf4484ef9034190b304d2f6b1