profile::base: Pin linux-base's version for Bookworm bpo
Update the apt::package_from_bpo define to allow passing
package versions. This allows puppet to install linux-base
to a specific version, rather than relying on the one already
installed by the OS. The linux-base package at version 4.2
is a pre-requisite to allow linux-image 6.12 to be installed
on Bookworm (from bpo).

Bug: T393948
Change-Id: Iabcf5e88dd54cbde5a08fd779a5e37f49b9ac774