Use the standby analytics_meta mariadb server temporarily
We are swapping the roles of an-mariadb100[1-2] temporarily, in order to
facilitate a firmware upgrade of the SSDs on an-mariadb1001.

This change will cause Hive and Druid to use the new hostname. It will
be reverted when the roles are reverted back on September 3rd.

Hosts: O:analytics_cluster::coordinator
Hosts: O:druid::analytics::worker
Bug: T394498
Change-Id: I9e90e98adc07a92cf0a69bebb36cd4d0373d6b26