Add a new insetup role for ml-k8s hosts to test their GPU
A lot of ml-k8s hosts are going to be equipped with GPUs,
and sometimes we need to reboot them after installing some packages
(like AMD firmware ones) to be able to get the hardware
recognized properly.

Add a new role for ml-k8s worker nodes, to be able to configure
a GPU before adding any k8s config/packages.

Apply the config to ml-serve1012 and ml-serve1013.

Bug: T393948
Change-Id: I826e6fc91785e689914df47dd0ef6b0e705b5f4b