replica_cnf: disable ssl by default on replica.cnf files
As we don't really support it.

Note that it's enabled in the override file (installed in the
bastaions) but also in the replica.my.cnf file for each user, as that
one is used inside jobs, that don't have the override file.

Also note that this will not change existing replica.my.cnf files,
only new ones.

Bug: T182892
Bug: T401861
Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: If80ab823c3d1f54f263d2ab71c881c93dc8055d2