dumps: Sync cirrus index dumps from hdfs
These dumps are to replace the existing cirrussearch dumps that
come from php. The previous dumps now take more than a week to
generate the weekly dumps. These replacement dumps use the hadoop
cluster and should have a latency of under one day.

These dumps are very similar, but the file layout is not exactly
the same. As such they are being placed in a new directory on the
dumps host. Some weeks after announcing availability of the new
dumps the old ones will be disabled.

Dumps orchestration is performed by the search platform airflow
instance.

Bug: T366248
Change-Id: I4132f271fe41ec33b9bcc6ec77defa6b6d1c8e78