zuul::main: use profile docker::engine to install docker
Instead of directly installing the docker.io package, use our
in-house profile to install docker.

We want a more standard installation and be able to use Hiera
keys to set parameters, specifically whether it uses iptables
or not.

Also include profile::docker::ferm (comments in there explain
why it's not included directly in docker::engine).

Currently the problem is that nftables is installed and not
iptables but docker wants to use iptables to map ports from container
to ports on the host.

Bug: T395938
Change-Id: Ia1f9da70e710caf12f2e2bed851d41915d16b05f