hieradata: add mw-next-routing to ATS tslua plugin chains
This adds the mw-next-routing ATS Lua script at the appropriate stage in
the mw-api-ext and mw-web tslua plugin chains.

Here, "appropriate stage" is immediately after multi-dc.lua, since the
logic in mw-next-routing.lua operates on -ro remapped target hosts.

The mw-next-routing script is a rename and lightly generalized version
of the mw-php-migration script originally used in the PHP 8.1 migration.
See T377042 for the original development of the latter and T391421 for
the rename / generalization.

This will be used again during the upcoming PHP 8.3 migration, but as of
now is a noop (i.e., remaps nothing) given the current configuration in
the mw-next-routing.lua.conf script.

Hosts: O:cache::text
Bug: T403655
Change-Id: Id74341233348a39a774e1eb222e89ab58605abb5